// src/models/ICustomer.ts
export default interface IAuthenticatedUser {
  name: string;
  displayName?: string;
  email: string;
  phone: string;
  metadata?: any;
  methods?: any;
  address?: UserAddress;
  origin?: string;
  memberId: string;
  default_source?: string;
  invoice_settings?: {
    default_payment_method: string | null;
  }
  id: string;
  first_name?: string;
  last_name?: string;
  iat: number;
  exp: number;
  fresh: number;
  exp_status: 'expired' | 'authed' | 'expiring';
  memberStatus: string;
  memberType: string;
}

export interface UserAddress {
  line1?: string;
  line2?: string;
  city?: string;
  state?: string;
  postal_code?: string;
  zip?: string;
  country?: string;
}